+++ WebFormHTR.Solution/WebFormHTR.Application/Features/Template/DTOs/TemplateWithoutParentDto.cs
@@ -5,8 +5,8 @@ namespace WebFormHTR.Application.Features.Template.DTOs;
 public record TemplateWithoutParentDto(
     Guid Id,
     string Name,
-    float Width,
-    float Height,
+    int Width,
+    int Height,
     FileDto? File,
     DateTime CreatedAt,
     DateTime UpdatedAt
