+++ WebFormHTR.Solution/WebFormHTR.Domain/ValueObjects/Coordinates.cs
@@ -2,8 +2,8 @@ namespace WebFormHTR.Domain.ValueObjects;
 
 public class Coordinates
 {
-    public float X { get; set; }
-    public float Y { get; set; }
-    public float Width { get; set; }
-    public float Height { get; set; }
+    public int X { get; set; }
+    public int Y { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
 }
