+++ WebFormHTR.Solution/WebFormHTR.Infrastructure/Services/Scripting/DTOs/PythonTemplateConfig.cs
@@ -7,6 +7,6 @@ public class PythonTemplateConfig
     [JsonPropertyName("content")] public IEnumerable<PythonRoiDto> Rois { get; set; } = [];
     [JsonPropertyName("to_ignore")] public IEnumerable<PythonResidualDto> Residuals { get; set; } = [];
 
-    [JsonPropertyName("width")] public float Width { get; set; }
-    [JsonPropertyName("height")] public float Height { get; set; }
+    [JsonPropertyName("width")] public int Width { get; set; }
+    [JsonPropertyName("height")] public int Height { get; set; }
 }
